% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_png_paths}
\alias{handle_png_paths}
\title{Handle file paths for PNG output}
\usage{
handle_png_paths(x, file_paths = NULL)
}
\arguments{
\item{x}{Character vector of names}

\item{file_paths}{Character string specifying the path(s) where to save the PNG file(s)}
}
\value{
Character vector of file paths
}
\description{
Handle file paths for PNG output
}
\keyword{internal}
